Better process management; support HUP signal, monitor tmp/restart.txt

+++ apps/worker/worker.rb
@@ -33,17 +33,21 @@ module Spider
         end
         
         def self.app_shutdown
-            return unless Spider.conf.get('worker.enable')
+            return unless @runner || Spider.conf.get('worker.enable')
             @mutex.try_lock || return
             Spider::Logger.info("Shutting down worker in #{Process.pid}")
+            Spider::Logger.debug("Runner: #{@runner}")
+            Spider::Logger.debug("PID FILE: #{File.exist?(@pid_file)}")
             if @runner
                 unless @runner == Thread.current
                     @runner.stop
                     @runner = nil
                 end
-            elsif(File.exist?(@pid_file))
+            elsif File.exist?(@pid_file)
                 begin
+                    puts "uno"
                     pid = IO.read(@pid_file).to_i
+                    puts "PID: #{pid}"
                     unless pid == Process.pid
                         Spider::Logger.info("Sending TERM signal to worker #{pid}")
                         Process::kill 'TERM', pid
@@ -99,14 +103,25 @@ module Spider
                 end
             end
             if (options[:fork])
-                forked = Spider.fork do
+                Spider.logger.debug("Forking worker in #{Process.pid}")
+                @forked = Spider.fork do
+                    $SPIDER_PROC_NAME ||= $0
                     $0 = 'spider-worker'
-                    trap('TERM') { app_shutdown }
-                    trap('INT') { app_shutdown }
+                    $SPIDER_NO_RESPAWN = true if $SPIDER_SPAWNED
+                    Spider.main_process_startup
+                    Spider.on_main_process_shutdown do
+                        # if $SPIDER_SPAWNED
+                        #     $SPIDER_PROC_NAME = 'spider'
+                        #     $SPIDER_PROC_ARGS = ['worker', 'start', '-d']
+                        # end
+                        Worker.app_shutdown
+                    end
+                    # trap('TERM') { app_shutdown }
+                    # trap('INT') { app_shutdown }
                     start.call
                     @runner.join if @runner
                 end
-                Process.detach(forked) if (options[:detach])
+                Process.detach(@forked) if (options[:detach])
                 @runner = nil
             else
                 start.call
